Second part of CR suggestions done
